fix: /code-reviewの指摘2件を修正(reps intでパース全滅・segments整合性ログ漏れ)
- MenuSegment.RepsをintのままにするとGeminiが"reps": 1.0のような小数表記で
  返した場合にjson.Unmarshalが失敗し、segments表示どころかアドバイス生成
  全体(GET /advices/latest)が失敗する経路だったため、他の数値フィールド同様
  float64に変更した
- ADR 022で「segmentsの不整合はmenu_type同様ログにのみ残す」と明記していたのに
  実装が漏れていたため追加: ビルドアップ走なのにsegments無し/他種別なのに
  segments有り/距離合計がdistance_kmと大きく食い違う、の3パターンをログに残す

+++ backend/internal/model/advice.go
@@ -57,8 +57,11 @@ func IsValidMenuType(s string) bool {
 // MenuSegment は1回の練習内で区間ごとに条件が変わる練習(現時点ではビルドアップ走のみ)の
 // 区間1つ分(docs/adr/022)。インターバル・変化走・ウェーブ走にも再利用できるよう、
 // RepsとRestSecを持たせているが、ビルドアップ走では常にReps=1・RestSec=0で使う。
+// Repsは意味上は整数(本数)だが、AIの出力は他の数値フィールド同様1.0のような
+// 小数表記で返ってくる可能性があり、int型にすると1件のパース失敗が
+// アドバイス生成全体を失敗させてしまう(/code-review指摘)ためfloat64にしている。
 type MenuSegment struct {
-	Reps         int     `json:"reps"`
+	Reps         float64 `json:"reps"`
 	DistanceKm   float64 `json:"distance_km"`
 	PaceSecPerKm float64 `json:"pace_sec_per_km"`
 	RestSec      float64 `json:"rest_sec"`
